update GDAL to 3.9.0 (#164)
* upgrade GDAL to 3.9.0

* seems that GDAL is C++17 now

* move C++17 to common

* use a more generic error message pattern

* try building with g++10

* a node-gyp special

* use the same C++ standard everywhere to avoid surprises with conditional compilation in the std lib

* it seems that GDAL 3.9 always returns an extent

* construct the metadata dynamically

* change this test to use 4326

* set the grib resource dir path

* use GNU C++17

* no need for g++-10

* fix the shell syntax

* more syntax

* missing comma

* this is a C++ file

* more C++ files

* remove all other std options

* these sources are not built in GDAL 3.9 anymore

* convert all spaces to tabs in gyp files (path of least changes)

* Set explicitly _HAS_EXCEPTIONS for the whole project

* this test is known to be flaky

+++ deps/libgdal/gdal/frmts/grib/gribcreatecopy.cpp
@@ -187,6 +187,7 @@ class GRIB2Section3Writer
 
   public:
     GRIB2Section3Writer(VSILFILE *fpIn, GDALDataset *poSrcDSIn);
+
     inline int SplitAndSwap() const
     {
         return nSplitAndSwapColumn;
@@ -1412,7 +1413,8 @@ bool GRIB2Section567Writer::WriteIEEE(GDALProgressFunc pfnProgress,
     WriteByte(m_fp, GRIB2MISSING_u1);  // no bitmap
 
     // Section 7: Data Section
-    const size_t nBufferSize = m_nXSize * GDALGetDataTypeSizeBytes(eReqDT);
+    const size_t nBufferSize =
+        static_cast<size_t>(m_nXSize) * GDALGetDataTypeSizeBytes(eReqDT);
     // section size
     WriteUInt32(m_fp, static_cast<GUInt32>(5 + nBufferSize * m_nYSize));
     WriteByte(m_fp, 7);  // section number
